dont allow unsetting sequenceViewer extension on backend close #1189
