+++ server/extensions/constants.js
@@ -1,2 +1,5 @@
-//route client requests for the extension bundle (e.g. extensions/load/extensionName/index.js)
-export const clientBundleUrl = 'index.js';
+export const sequenceViewerExtension = 'GC-Sequence-Viewer';
+
+export const fixedExtensions = {
+  [sequenceViewerExtension]: true,
+};
