+++ server/extensions/router.js
@@ -19,14 +19,13 @@ import express from 'express';
 import { ensureReqUserMiddleware } from '../user/utils';
 import errorHandlingMiddleware from '../utils/errorHandlingMiddleware';
 import { errorDoesNotExist } from '../utils/errors';
-import extensionApiRouter from './apiRouter';
-import { getExtensionInternalPath } from './loadExtension';
+import extensionApiRouter from './routerApi';
 import { manifestIsClient, manifestIsServer } from './manifestUtils';
-import { checkClientExtensionFilePath, checkExtensionExistsMiddleware, checkExtensionIsClientMiddleware, checkUserExtensionAccess, checkUserExtensionAccessMiddleware, checkUserExtensionActive } from './middlewareChecks';
+import { checkClientExtensionFilePath, checkExtensionExistsMiddleware, checkExtensionIsClientMiddleware, checkUserExtensionAccess, checkUserExtensionAccessMiddleware, checkUserExtensionActive } from './routerMiddleware';
 import csvRouter from './native/csv/index';
 import fastaRouter from './native/fasta/index';
 import genbankRouter from './native/genbank/index';
-import { getExtensions } from './registry';
+import { getExtensionInternalPath, getExtensions } from './registry';
 
 const router = express.Router(); //eslint-disable-line new-cap
 const jsonParser = bodyParser.json();
