+++ server/user/updateUserHandler.js
@@ -17,9 +17,11 @@
 import debug from 'debug';
 import EmailValidator from 'email-validator';
 import fetch from 'isomorphic-fetch';
+import _ from 'lodash';
 
 import { headersPost } from '../../src/middleware/utils/headers';
 import userConfigDefaults from '../onboarding/userConfigDefaults';
+import userConfigOverrides from '../onboarding/userConfigOverrides';
 import { verifyCaptchaProductionOnly } from '../utils/captcha';
 import { API_END_POINT, INTERNAL_HOST } from '../urlConstants';
 import { mergeConfigToUserData, pruneUserObject, updateUserAll, updateUserConfig, validateConfig } from './utils';
@@ -62,7 +64,9 @@ export function registrationHandler(req, res, next) {
     }
   }
 
+  //shallow assign, so explicitly declare projects + extensions, not merging with defaults
   const mergedConfig = Object.assign({}, userConfigDefaults, config);
+  _.merge(mergedConfig, userConfigOverrides);
 
   try {
     validateConfig(mergedConfig);
